Set
initial API for the Set CRDT. Operates similarly to an Observed Remove
Set - in that each element is tagged with a Dot ({actor, clock} tuple).
Removing an element deletes it locally, and should record the Dot so it
can be deleted on remote replicas.

This set is a named CRDT - each replica operates on a separate
"partition" of the set. We can make this assumption because it is
intended for solving a problem like Phoenix.Presence - where each node
is tracking PIDs that are local to that node, no other node will operate
on those PIDs.

Next step is adding Deltas - a store of additions/deletions that can
easily be sent to other replicas and merged.